kscope: adblock: Allow overriding the input hosts file
Because of the modular nature of the StevenBlack's hosts, the user may be tempted to modify the hosts to match personal preferences. Accommodate such changes.
Retains the earlier behaviour to pass multiple files to be conbined.

Also remove the shell script "middleman", earlier shell script was used to fetch the file versions, which now is handled within the python script itself

Change-Id: I4a348f29c9dbec6eb9eafe387a14be7280ce2412